Rollup of small fixes

+++ src/lib/modes/aead/chacha20poly1305/chacha20poly1305.h
@@ -56,11 +56,11 @@ class ChaCha20Poly1305_Mode : public AEAD_Mode {
 
       secure_vector<uint8_t> m_ad;  // NOLINT(*non-private-member-variable*)
       size_t m_nonce_len = 0;       // NOLINT(*non-private-member-variable*)
-      size_t m_ctext_len = 0;       // NOLINT(*non-private-member-variable*)
+      uint64_t m_ctext_len = 0;     // NOLINT(*non-private-member-variable*)
 
       bool cfrg_version() const { return m_nonce_len == 12 || m_nonce_len == 24; }
 
-      void update_len(size_t len);
+      void update_len(uint64_t len);
 
    private:
       void start_msg(const uint8_t nonce[], size_t nonce_len) override;
